+++ Courier/Changes.md
@@ -3,6 +3,7 @@
 ## v1.1.3 (in development)
 ### Bugfixes
 - Added support for nested messages
+- Added handling of `null` message subject values
 
 ## v1.1.2 - 04.10.2024
 A new build with Scala v2.13.14. No other changes.